Changes to support netstandard for netcore and dual targeting to net45
Removed WebClient references to adapt to a more .net core way with
HttpClient
Migrated away from explicit async methods in QueryAsync and instead made
Query async
CancelationTokens are no longer required on the API, and instead are
passed in via tasks which support cancelation tokens in them and will be
more familar to API consumers.
Updated unit tests to work against new API
Updated travis build script to support building dotnetcore dual targets
on Ubuntu
Restructured project to meet netcore best practices (ie src / test)
Added an artifacts folder so that all items can be seen in Visual Studio
Removed the connection pooling as this can be done by the consumer via
tasks if required. It could be added back as a future enhancement. It
would be adventageous to do this at a task level since that is more
abstract and could handle non http approaches in the future.