Removed webclient library dependency
MapsAPIGenericEngine

removed obsolete methods that are private and protected but without reference elsewhere to simplify its update away from webclient to httpclient. Marked the methods in here as obsolete as they are superfluous now because of the switch to httpclient which exposes the same async behavior out of box.

EngineFacade

Marked methods obsolete as .net handles performance better with async that it does using asyncrous task creation because of the way tools like IIS will pool threads with server connections and free them up, traditional tasks don't do that and hence shouldn't be used. Methods kept for API compatability

HttpClientExtensions (Added)

Helper methods for backwards compatiability of API. Marked class and methods obsolete as they don't add value over the basic HttpClient class which offers the same behaviors out of box.

Project.json (added)

Somehow this was required by VS 2017 to get the project to work.

WebclientEx (removed)

All functions of this are on httpclient by default.